Only pull in the method we override
